migrate to lualine (from galaxyline)
galaxyline was annoying me for a while (had to have my fork to be able
to display inactive buffer status lines, generally unmaintained). now i
also noticed that galaxyline was doing A LOT of stats on the disk
(basically at every cursor movement), which was killing the performance
when working over sshfs. lualine fixes that. and is maintained.